requirements.txt
python-telegram-bot==21.5
openai==1.45.0
python-dotenv==1.0.1
elevenlabs==1.8.1
replicate==0.33.0
pydub==0.25.1